Collision detection added.
Walls are modeled as a list of pixels which leads to weird jittering
see bellow and collision detection works by running through all the
potential pixels that could be collided with.

TODO: There is a weird jittering when running into walls head on, this
is because we are modelling walls as a collection of pixels and on
collision we rapidly swap between coliding with 2 different pixels.

This shouldn't be noticeable with the controller as I believe there will
almost always be a slight glide one way or the other.

Fix: When line collisions are implemented this shouldn't happen any
more.

TODO: Work out a way of reducing the number of pixels to collide with.

TODO: At present the is a slight overlap when colliding with the pixels
above and to the left. This is because we are colliding with the top
right corner of the pixel.